omega/triage-portal/src/requirements.txt
asgiref==3.4.1
astroid==2.9.0
azure-core==1.21.1
azure-storage-blob==12.9.0
bandit==1.7.1
black==21.11b1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.0
Deprecated==1.2.13
Django==4.0.1
django-dotenv==1.4.2
django-redis==5.1.0
django-taggit==2.0.0
dodgy==0.2.1
flake8==4.0.1
flake8-polyfill==1.0.2
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
isodate==0.6.0
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
msrest==0.6.21
mypy-extensions==0.4.3
oauthlib==3.1.1
packageurl-python==0.9.6
pathspec==0.9.0
pbr==5.8.0
pep8-naming==0.12.1
platformdirs==2.4.0
prospector==0.12.2
psycopg2==2.9.2
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
pylint==2.12.2
pylint-celery==0.3
pylint-common==0.2.5
pylint-django==2.4.4
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0
redis==4.0.2
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
requirements-detector==0.7
setoptconf==0.3.0
setoptconf-tmp==0.3.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sqlparse==0.4.2
stevedore==3.5.0
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.1
urllib3==1.26.7
wrapt==1.13.3
